% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under Curve}
\usage{
auc(x, t = NULL, dt = NULL, na.rm = T)
}
\arguments{
\item{x}{data values}

\item{t}{time values (x axis)}

\item{dt}{optional data.table}

\item{na.rm}{remove missing values}
}
\value{
total area bounded by the points (t, x)
}
\description{
Estimates the area under a given curve using the trapezoidal rule, which is equivalent to the average of a left and right Riemann sum.
}
