% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{deltaf}
\alias{deltaf}
\title{Delta F over F}
\usage{
deltaf(x, r, fun = mean)
}
\arguments{
\item{x}{values}

\item{r}{baseline values}

\item{fun}{summary function (mean)}
}
\value{
percent departure from baseline
}
\description{
Delta F over F
}
