% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiPhA.R
\name{FiPhA}
\alias{FiPhA}
\title{Starts FiPhA's Shiny Application}
\usage{
FiPhA(...)
}
\description{
Starts the Shiny user interface application.
}
\examples{
# start in a browser
# FiPhA::FiPhA(launch.browser = T)

}
