% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{pctdff}
\alias{pctdff}
\title{Percent Delta F over F}
\usage{
pctdff(x, r, fun = mean)
}
\arguments{
\item{x}{values}

\item{r}{baseline values}

\item{fun}{summary function (mean)}
}
\value{
delta f over f expressed as a percent
}
\description{
Percent Delta F over F
}
