% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt.R
\name{read_tev}
\alias{read_tev}
\title{Read raw values contained in tev file using tsq data}
\usage{
read_tev(
  path,
  tsq.headers = NULL,
  streams = c("405A", "465A", "560B"),
  wide = T
)
}
\arguments{
\item{path}{folder containing TDT data files (*.tsq, *.tev, etc.)}

\item{tsq.headers}{tsq header information from parse_tsq}

\item{streams}{which data streams to parse?}

\item{wide}{return data in a wide format rather than long}

\item{tsq.file}{tsq file name}

\item{tev.file}{tev file name}
}
\value{
data.table of raw values
}
\description{
Given a set of event headers found by reading the tsq file, read data into usable time series formats.

TODO: fix for importing multiple frequencies at once
}
