% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt.R
\name{read_tsq}
\alias{read_tsq}
\title{Open and parse a *.tsq file's data block and event metadata}
\usage{
read_tsq(path)
}
\arguments{
\item{path}{folder containing TDT data files}
}
\value{
data.table containing metadata for the associated .tev file
}
\description{
Open and parse a *.tsq file's data block and event metadata
}
