% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{robustz}
\alias{robustz}
\title{Robust Z}
\usage{
robustz(x, r, fun = mean)
}
\arguments{
\item{x}{values}

\item{r}{baseline values}

\item{fun}{summary function (mean)}
}
\value{
zscore-like values based on medians
}
\description{
Robust Z
}
