% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt.R
\name{tsq_info}
\alias{tsq_info}
\title{Print information about the available data streams}
\usage{
tsq_info(tsq.headers)
}
\arguments{
\item{tsq.headers}{metadata from read_tsq}

\item{path}{folder containing TDT data files}

\item{tsq.file}{tsq file name}
}
\description{
Print information about the available data streams
}
