% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiPhA.R
\name{update_FiPhA}
\alias{update_FiPhA}
\title{Update FiPhA}
\usage{
update_FiPhA(force.update = F, upgrade.dependencies = F)
}
\arguments{
\item{force.update}{upgrade even if already the latest version}

\item{upgrade.dependencies}{upgrade dependencies if available}
}
\description{
Update the package by re-installing the latest build from its github branch, optionally updating packages along the way. Skips updating by default if
the package was installed locally.
}
