% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{zscore}
\alias{zscore}
\title{Zscore}
\usage{
zscore(x, r = x, fun = mean)
}
\arguments{
\item{x}{values}

\item{r}{baseline values (x)}

\item{fun}{summary function (mean)}
}
\value{
standard z score
}
\description{
Zscore
}
