% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FiPhA.R
\name{update_FiPhA}
\alias{update_FiPhA}
\title{Update FiPhA from its GitHub repo}
\usage{
update_FiPhA(force.update = F, upgrade.dependencies = F)
}
\arguments{
\item{force.update}{upgrade even if already the latest version}

\item{upgrade.dependencies}{upgrade dependencies if available}
}
\description{
Update FiPhA from its GitHub repo
}
